pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "expenses-analyzer"
version = "0.1.0"
authors = [
  { name = "Angelo 'pallotron' Failla", email = "pallotron@gmail.com" },
]
description = "A Textual TUI application for analyzing your personal expenses."
readme = "README.md"
requires-python = ">=3.12"
classifiers = [
  "Programming Language :: Python :: 3",
  "Operating System :: OS Independent",
]
dependencies = ["pandas", "textual", "pyarrow", "google-generativeai"]

[project.optional-dependencies]
dev = ["pytest-asyncio"]

[project.scripts]
expenses-analyzer = "expenses.main:main"

[tool.setuptools]
packages = {find = {}}
include-package-data = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"

[tool.coverage.run]
source = ["expenses"]
omit = [
    "*/tests/*",
    "*/__pycache__/*",
    "*/venv/*",
    "*/.venv/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "@abstractmethod",
]
precision = 2
show_missing = true

[tool.coverage.html]
directory = "htmlcov"
